+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `load_all()` now issues a warning if exported objects conflict with objects defined in the global environment (#112)
 
+* `dev_example()` now works after removing an inconsistent call to `load_all()` (@riccardoporreca, #122).
+
 # pkgload 1.0.2
 
 * `shim_question()` now works for topics from the R base package that are passed with the double colon operator (e.g. `base::min`) (@mdequeljoe, #99).