+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * `dev_example()` now works after removing an inconsistent call to `load_all()` (@riccardoporreca, #122).
 
+* Internal fixes for compatibility with the future 4.1.0 release.
+
 # pkgload 1.0.2
 
 * `shim_question()` now works for topics from the R base package that are passed with the double colon operator (e.g. `base::min`) (@mdequeljoe, #99).